services/llm_gateway/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
google-generativeai
python-dotenv
